Added controls for new hardware & position constants
Moved the carousel controls to the trigger to help ramp the speed with the analog value. (Deadzone may need to be reduced)
Basic states for arm set-positions. The arm naturally slides down because of gravity and the elastic tubing, meaning the motor must constantly be powering it. The manual state does not have it holding a position, this needs to be fixed.