I hate everything
By working with the contraption known as the 'Intel Realsense T265', I have lost my already waning will to live. My future prospects towards life have diminished to the point of living in a plastic box for the next decades of my life is comparatively better than working with this contraption. This is a cry for help. I have lost my remaining thread of sanity. Please send help now.