Working autonomous, name changes, trajectory changes, arm position changes
Renamed the StateTypes from last year in order to make more sense when reading code. 

Made some major trajectory changes:
- Working red carousel, previously had some issues with the rotation being wrong.
- Attempt at a warehouse starting point, blue only currently.
- Experimenting with going into the warehouse to grab a second cargo element and scoring that, then parking in the warehouse. These trajectories are only theoretical. Blue only.

New arm positions for the bottom & middle positions of the shipping hub. Since we have an additional servo that acts as a lid, it will get stuck on the robot and prevent the drop from happening. This is fixed by reducing the amount of time spent on the horizontal linear slide and increasing the height of the vertical linear slide.

Major state machine changes:
- Added new logic for the vision on the warehouse side. Because the robot cannot start with seeing the placement square, we have to use the middle & top positions. Red is going to be the same, but that's a problem for later.
- New states for the warehouse & second-cargo autonomous.
- States utilize all of the servos on the cargo basket. 
- Added a few TODOs to convert plain numbers into variables, allowing for easier editing & reading.